Fix(AddPoll): Separate logic for polls and fix listing options
