+++ ClubestApp/Services/ClubService.cs
@@ -173,107 +173,6 @@
 
             return result.Entity;
         }
-
-        public async Task<Poll> AddVote(List<string> votes, string pollId, string userId)
-        {
-            User user = await this.dbContext.Users.FirstOrDefaultAsync(u => u.Id == userId);
-            Poll currentPoll = null;
-
-            foreach (var vote in votes)
-            {
-                Option option = await this.dbContext
-                    .Options
-                    .FirstOrDefaultAsync(x => x.Content == vote && x.PollId == pollId);
-                
-                currentPoll = await this.dbContext
-                    .Polls
-                    .FirstOrDefaultAsync(x => x.Id == pollId);
-                if (option != null && currentPoll.PollVotedUsers.Any(x => x.UserId == userId) == false)
-                {
-                    option.VotesCount++;
-                    currentPoll.PollVotedUsers.Add(new PollVotedUsers
-                    {
-                        User = user,
-                        UserId = userId,
-                        PollId = pollId,
-                        Poll = currentPoll
-                    });
-
-                    await this.dbContext.SaveChangesAsync();
-                }
-
-            }
-
-            return currentPoll;
-        }
-
-        public async Task<Poll> CreatePoll(AddPollInputModel model, string id)
-        {
-            Poll newPoll = new Poll()
-            {
-                Content = model.Content,
-                ClubId = id,
-                IsMultichoice = model.IsMultichoice.ToLower() == "true" ? true
-                : false,
-                ExpiredDate = model.ExpiredDate,
-            };
-
-            var poll = await this.dbContext.Polls.AddAsync(newPoll);
-
-            List<string> options = model.Options.Split(" ", StringSplitOptions.RemoveEmptyEntries)
-                .ToList();
-
-            foreach (var option in options)
-            {
-                Option newOption = new Option()
-                {
-                    VotesCount = 0,
-                    Content = option,
-                    PollId = poll.Entity.Id,
-                };
-
-                await this.dbContext.Options.AddAsync(newOption);
-            }
-
-            await this.dbContext.SaveChangesAsync();
-            return poll.Entity;
-        }
-
-        public async Task<ListPollsBindingModel> GetPollsModel(string clubId, string userId)
-        {
-            Club club = await this.dbContext
-                .Clubs
-                .FirstOrDefaultAsync(x => x.Id == clubId);
-
-            ListPollsBindingModel result = new ListPollsBindingModel
-            {
-                ClubPriceType = club.PriceType.ToString(),
-                Club = club,
-                Polls = this.dbContext
-                        .Polls
-                        .Where(x => x.ClubId == clubId && this.IsPollValid(x.ExpiredDate))
-                        .Select(x => new PollItemBindingModel
-                        {
-                            Id = x.Id,
-                            Content = x.Content,
-                            VotesCount = x.PollVotedUsers.Count,
-                            ExpiredDate = x.ExpiredDate,
-                            Options = x.Options.ToList(),
-                            IsMultichoice = x.IsMultichoice,
-                            PollVotedUsers = x.PollVotedUsers.ToList()
-                        })
-                        .ToList(),
-                ClubId = club.Id,
-                UserId = this.dbContext.Users.First(u => u.Id == userId).Id
-            };
-
-            return result;
-        }
-
-        private bool IsPollValid(DateTime expiredDate)
-        {
-            return expiredDate.Subtract(DateTime.UtcNow).Hours > 0;
-        }
       
         public async Task<Club> EditClub(EditClubInputModel model, string id)
         {
