+++ ClubestApp/Views/Shared/_AddPoll.cshtml
@@ -11,7 +11,7 @@
                 </button>
             </div>
 
-            <form asp-controller="Club" asp-action="AddPoll" method="post">
+            <form asp-controller="Poll" asp-action="AddPoll" method="post">
                 <div class="pt-1 modal-body mx-3 ">
                     <p>Вид</p>
                     <div class="custom-control custom-radio  pr-2 mx-auto" style="display: inline-block">
@@ -35,7 +35,7 @@
 
                     <div class="md-form">
                         <input type="text" id="optionField" class="form-control validate">
-                        <label data-error="wrong" data-success="right" for="form34">Добави опция</label>
+                        <label for="form34">Добави опция</label>
                     </div>
 
                     <button type="button" class="btn aqua-gradient px-3 mb-2" id="addButton">Добави</button>
